fix: bottoni snooze non funzionanti - 3 cause risolte
1. AlarmModal: canSnooze calcolato con getSnoozeDelay()=0 evita
   falsi negativi quando l'allarme scatta esattamente al minuto esatto
2. App.tsx: setAlarmingScheduleId(null) spostato PRIMA della chiamata
   bridge Android, con try/catch; il modal si chiude sempre
3. WebAppInterface: delayMs Long → Double perché il WebView bridge
   mappa i numeri JS a double Java; con Long il metodo non veniva trovato
   e la chiamata falliva silenziosamente

+++ android/app/src/main/java/it/memofarmaci/app/WebAppInterface.kt
@@ -23,9 +23,11 @@ class WebAppInterface(private val activity: MainActivity) {
     @JavascriptInterface
     fun getAlarmPlanId(): String = activity.getAlarmPlanId()
 
-    /** Schedula un allarme nativo (snooze) via AlarmManager */
+    /** Schedula un allarme nativo (snooze) via AlarmManager.
+     *  delayMs è Double perché JavaScript non ha Long e il WebView bridge
+     *  mappa i numeri JS a double Java: se usiamo Long il metodo non viene trovato. */
     @JavascriptInterface
-    fun scheduleSnoozeAlarm(planId: String, delayMs: Long) {
-        activity.scheduleSnoozeAlarm(planId, delayMs)
+    fun scheduleSnoozeAlarm(planId: String, delayMs: Double) {
+        activity.scheduleSnoozeAlarm(planId, delayMs.toLong())
     }
 }
